updated git.sh

+++ git.sh
@@ -1,2 +1,9 @@
-git fetch --all
-git reset --hard origin/master
+#!/bin/bash
+echo This will overwrite any local changes you have made.
+echo Continue? (y/n)
+read answer
+if [ "$answer" -eq "y" ]
+then
+	git fetch --all
+	git reset --hard origin/master
+fi
